Update SoftCam.Key
Exclusively for Enigma2 and OsCam

Thanks to everyone who sent the keys with Namespace Hash, because my rotor is broken! Now PowerVu with new format, AU around 2-3 minutes. If something is missing, write me, we will correct and add!

Merry Christmas and good luck to everyone in the NEW 2020 Rat Year!!!